BuildTile -> BuildCore
